requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==10.4.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
